# Consider memory during writer scaling