… core/trino-main/src/main/java/io/trino/sql/planner/LocalExecutionPlanner.java
… core/trino-main/src/test/java/io/trino/execution/TestTaskManagerConfig.java
… core/trino-main/src/test/java/io/trino/operator/exchange/TestLocalExchange.java
… core/trino-main/src/test/java/io/trino/operator/join/unspilled/JoinTestUtils.java
… core/trino-main/src/test/java/io/trino/operator/output/TestSkewedPartitionRebalancer.java
… core/trino-main/src/test/java/io/trino/sql/analyzer/TestFeaturesConfig.java
… plugin/trino-hive/src/test/java/io/trino/plugin/hive/BaseHiveConnectorTest.java
… testing/trino-faulttolerant-tests/src/test/java/io/trino/faulttolerant/hive/TestHiveFaultTolerantExecutionConnectorTest.java