<!-- Thank you for submitting a pull request! Find more information in our development guide at https://github.com/trinodb/trino/blob/master/.github/DEVELOPMENT.md and contact us on #dev in Slack. -->
<!-- Provide an overview of the PR for maintainers and reviewers. -->
## Description

## Benchmarks (Unpartitioned)

Before:
```
trino> set session query_max_memory_per_node = '3GB';
SET SESSION
trino:insert_demo> insert into lineitem_unpart_2 select * from tpch_sf1000_orc.lineitem;

Query 20230829_001613_00003_7gygv, FAILED, 7 nodes
Splits: 903 total, 16 done (1.77%)
11.32 [36.1M rows, 13.1GB] [3.19M rows/s, 1.16GB/s]

Query 20230829_001613_00003_7gygv failed: Query exceeded per-node memory limit of 3GB [Allocated: 3.00GB, Delta: 1010.93kB, Top Consumers: {TableScanOperator=2.51GB, TableWriterOperator=503.44MB, AggregationOperator=115.58kB}]
```

After:
```
trino> set session query_max_memory_per_node = '3GB';
SET SESSION
trino:insert_demo> insert into lineitem_unpart select * from tpch_sf1000_orc.lineitem;
INSERT: 5999989709 rows

Query 20230828_224641_00005_ftzqd, FINISHED, 7 nodes
Splits: 3,116 total, 3,116 done (100.00%)
8:52 [6B rows, 172GB] [11.3M rows/s, 330MB/s]
```

## Benchmarks (Partitioned)

Before:
```
trino:insert_demo> set session query_max_memory_per_node = '6GB';
SET SESSION
trino:insert_demo> insert into lineitem_few_skewed_partitions select orderkey, partkey, suppkey, linenumber, quant
ity, extendedprice, discount, tax, returnflag, linestatus, commitdate, receiptdate, shipinstruct, comment, shipdat
e, shipmode from tpch_sf1000_orc_part.lineitem;

Query 20230831_032052_00012_tus9r, FAILED, 7 nodes
Splits: 1,779 total, 48 done (2.70%)
21.84 [196M rows, 17.3GB] [8.97M rows/s, 811MB/s]

Query 20230831_032052_00012_tus9r failed: Query exceeded per-node memory limit of 6GB [Allocated: 6.00GB, Delta: 3.00MB, Top Consumers: {ScanFilterAndProjectOperator=3.18GB, TableWriterOperator=2.35GB, PagePartitioner=454.33MB}]
```


After:
```
trino:insert_demo> set session query_max_memory_per_node = '6GB';
SET SESSION
trino:insert_demo> insert into lineitem_few_skewed_partitions select orderkey, partkey, suppkey, linenumber, quant
ity, extendedprice, discount, tax, returnflag, linestatus, commitdate, receiptdate, shipinstruct, comment, shipdat
e, shipmode from tpch_sf1000_orc_part.lineitem;
INSERT: 5999989709 rows

Query 20230831_031245_00014_dhrst, FINISHED, 7 nodes
Splits: 5,463 total, 5,463 done (100.00%)
11:48 [6B rows, 159GB] [8.48M rows/s, 230MB/s]
```

_**However, note that this query takes ~4 minutes with 64GB query_max_memory_per_node.**_

<!-- Provide details that would help an engineer who is unfamiliar with this part of the code. -->
## Additional context and related issues



<!-- Mark the appropriate option with an (x). Propose a release note if you can. -->
## Release notes

( ) This is not user-visible or is docs only, and no release notes are required.
( ) Release notes are required. Please propose a release note for me.
( ) Release notes are required, with the following suggested text:

```markdown
# Section
* Fix some things. ({issue}`issuenumber`)
```