@sopel39 2023-09-25
Would be great to address https://github.com/trinodb/trino/pull/18771#discussion_r1335744423 as a follow-up